refactor: store all new components under one tests dir, simplified structure
