+++ src/App.js
@@ -1,7 +1,7 @@
 // dependencies
 import React from 'react';
 // local files
-import Button from './components/Button/Button';
+import Button from './components/Button';
 
 const App = () => {
   return (
